backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
apscheduler==3.10.4
email-validator==2.1.0
pyjwt
passlib[bcrypt]
python-jose    # optional if you prefer jose instead of pyjwt
bcrypt>=4.0.1
passlib>=1.7.4
